AbnormalReturns Data Structure

The key to the performance in this package is the underlying data structure. These rely on a combination of BusinessDays.jl and Tables.jl to provide fast access to slices of data based on dates.

DataVector (and DataMatrix)

AbnormalReturns.DataVector — Type

struct DataVector <: CalendarData
    data::Vector{Float64}
    missing_bdays::SparseVector{Bool, Int}
    dates::ClosedInterval{Date}
    calendar::MarketCalendar
    function DataVector(data, missing_bdays, dates, calendar)
        @assert length(data) == length(missing_bdays) == bdayscount(calendar, dates.left, dates.right) + 1 "Data does not match length of dates or missings"
        new(data, missing_bdays, dates, calendar)
    end
end

DataVector(data::AbstractVector, dates::ClosedInterval{Date}, cal::MarketCalendar)

DataVector(data::AbstractVector, dates::AbstractVector{Date}, cal::MarketCalendar)

Missing docstring.

Missing docstring for DataMatrix. Check Documenter's build log for details.

These structures provide strongly typed data that is easy to slice based on a range of dates. The data is always stored as Float64, even though it accepts elements of type Missing. In storing the data, Missing values are converted to 0.0, and the missing_bdays is a SparseVector that is true when that value is missing. dates are the minimum and maximum dates for the data.

Access to a DataVector or Matrix is through providing a ClosedInterval from IntervalSets.jl:

data[Date(2018) .. Date(2019)]

This will return a Vector or Matrix, depending on type.

!!! Note The data returned is always a Vector (or Matrix) of Float64, even if there was missing data in the original Vector. To truly get back to the original, you must combine the missing_bdays with the Vector.

MarketData

    struct MarketData{T, MNames, FNames, N1, N2}
        calendar::MarketCalendar
        marketdata::NamedTuple{MNames, NTuple{N1, DataVector}} # column names as symbols
        firmdata::Dict{T, NamedTuple{FNames, NTuple{N2, DataVector}}} # data stored by firm id and then by column name as symbol
    end

This struct is made up of a set of DataVector. The main purpose of this is to provide efficient storage of the underlying data.

TimelineTable

AbnormalReturns.TimelineTable — Type

mutable struct TimelineTable{Mssng, T, MNames, FNames, N1, N2} <: Tables.AbstractColumns
    parent::MarketData{T, MNames, FNames, N1, N2}
    "The current ID"
    id::T
    "Whether the functions return Vector{Union{Missing, Float64}} or Vector{Float64}"
    allow_missing::Type{AllowMissing{Mssng}}
    "Actual returnable dates that guarentees a square matrix made of the minimum and maximum of the DataVectors for this ID and set of columns"
    dates::ClosedInterval{Date}
    "Index of column names and any lags/leads"
    cols::DictIndex
    "missing days between the dates"
    missing_bdays::SparseVector{Bool, Int}
    "dates requested, which provides what data is automatically returned by functions"
    req_dates::ClosedInterval{Date}
end

This type provides Tables.jl access to DataVectors.

Functions to update some values are:

AbnormalReturns.update_id!: changes the id (which updates dates and missing_bdays)

select!: changes the cols (also updates dates and missing_bdays if cols is different)

AbnormalReturns.update_dates!: changes req_dates (does not update anything else)